Moved purely Address and Entry tests to dedicated test classes.
