+++ .gitignore
@@ -5,4 +5,5 @@ phpunit.xml
 vendor
 composer.lock
 build/logs
-
+build/
+/nbproject/*
